Added fluid background in hero
