+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "portfolio-frontend",
-  "version": "0.8.0",
+  "version": "0.9.0",
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
