+++ src/app/page.tsx
@@ -1,9 +1,10 @@
-import { Projects, Story, Services, SocialPosts } from '@/components';
+import { Hero, Projects, Story, Services, SocialPosts } from '@/components';
 
 export default function page() {
     return (
         <>
             <main className="main_flex">
+                <Hero />
                 <Projects />
                 <Story />
                 <Services />
