--- src/components/FluidShader/index.ts
@@ -1,5 +1,0 @@
-export { default as FluidShader } from './FluidShader';
-export * from './types';
-export * from './presets';
-export * from './utils';
-export { vertexShader, fluidShader, displayShader } from './shaders';
