+++ src/components/sections/index.ts
@@ -1,3 +1,5 @@
+export * from './Hero/Hero';
+
 export * from './Projects/Projects';
 
 export * from './Story/Story';
