+++ src/components/ui/FluidGradientBackground/FluidBackgroundPresets.ts
@@ -1,0 +1,11 @@
+export const DEFAULT_PRESET = {
+    colors: {
+        color1: '#8b96e8', // Deep blue
+        color2: '#335dea', // Purple-blue
+        color3: '#0d0d0d', // Bright blue
+        color4: '#0700c5', // Royal blue
+    },
+    colorIntensity: 1.2,
+    pixelRatio: 0.2,
+    fps: 60,
+};
