+++ src/components/ui/Noise/Noise.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import styles from '@/styles/components/ui/Noise/Noise.module.css';
+
+export const Noise = () => {
+    return <div className={styles.noise} />;
+};
