+++ src/components/ui/SectionSubHeading/SectionSubHeading.tsx
@@ -3,7 +3,8 @@
 import React, { useRef } from 'react';
 import gsap from 'gsap';
 import { useGSAP } from '@gsap/react';
-import { SplitText } from '@/components/utility/gsap';
+
+import { SplitText } from '@/utility/gsap';
 
 import { SectionSubHeadingProps } from './types';
