+++ src/components/ui/index.ts
@@ -5,3 +5,7 @@ export * from './SectionHeading/SectionHeading';
 export * from './SectionSubHeading/SectionSubHeading';
 
 export * from './SectionTextContent/SectionTextContent';
+
+export * from './FluidGradientBackground/FluidGradientBackground';
+
+export * from './Noise/Noise';
