+++ src/styles/components/sections/Hero/Hero.module.css
@@ -1,0 +1,6 @@
+.hero_container {
+    height: 100vh;
+    width: 100vw;
+    position: relative;
+    overflow: hidden;
+}
