+++ src/styles/components/ui/FluidGradientBackground/FluidGradientBackground.module.css
@@ -1,0 +1,7 @@
+.fluid_gradient_background_container {
+    height: 100%;
+    width: 100%;
+    top: 0;
+    left: 0;
+    z-index: -1;
+}
