+++ src/styles/components/ui/Noise/Noise.module.css
@@ -1,0 +1,8 @@
+.noise {
+  position: absolute;
+  height: 100vh;
+  width: 100vw;
+  top: 0;
+  left: 0;
+  background-image: url('/noise.svg');
+}
